requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg[binary]
python-dotenv
pydantic[email]
redis
python-multipart
python-jose[cryptography]
passlib[bcrypt]
PyYAML

# Testing
pytest
pytest-asyncio
httpx
